refactor : 변수명 통일 (#125)
