+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/product/infrastructure/ProductCategoryAdapter.java
@@ -16,9 +16,7 @@ public class ProductCategoryAdapter implements LoadProductCategoryPort {
 
     @Override
     public List<ProductCategory> findAllByIds(List<Long> productCategoryIds) {
-        List<ProductCategoryEntity> entities = jpaProductCategoryRepository.findAllById(productCategoryIds);
-
-        return entities.stream()
+        return jpaProductCategoryRepository.findAllById(productCategoryIds).stream()
                 .map(ProductCategoryEntity::toProductCategory)
                 .toList();
     }
