+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/product/presentation/response/ProductFavoriteDetailsResponse.java
@@ -10,7 +10,6 @@ import java.util.List;
 
 @Getter
 @NoArgsConstructor
-@AllArgsConstructor
 public class ProductFavoriteDetailsResponse {
     private Long productId;
     private Long bakeryId;
