+++ breaditnow-domain/src/main/resources/application-mysql-prod.yml
@@ -18,7 +18,7 @@ spring:
   sql:
     init:
       mode: never
-      data-locations: classpath:db/address.sql.sql, classpath:db/bread-category.sql
+      data-locations: classpath:db/address.sql, classpath:db/bread-category.sql
 
 geolocation:
   kakao:
